pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "githubagent"
version = "0.1.0"
description = "A professional GitHub development assistant agent powered by LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = ["github", "agent", "llm", "openai", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "PyGithub>=2.1.1",
    "openai>=1.14.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "requests>=2.31.0",
]

[project.scripts]
githubagent = "main:cli"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
include = ["prompts*", "tools*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
